how-computers-work videos repositioned (#27)
Moves the video-watching assignment to Class 6, as a second LJ assignment. 

This makes space in class 8 for reading about loops. But it also gives students a chance to learn about computers before really digging in to JavaScript. 

This is meant to be more of a homework assignment, rather than something completed during class time.